+++ src/providers/LinkResolver.cpp
@@ -32,12 +32,11 @@ void LinkResolver::getLinkInfo(
                 auto statusCode = root.value("status").toInt();
                 QString response = QString();
                 QString linkString = url;
-                ImagePtr thumbnail = nullptr;
+                ImagePtr thumbnail =
+                    Image::fromUrl({root.value("thumbnail").toString()});
                 if (statusCode == 200)
                 {
                     response = root.value("tooltip").toString();
-                    thumbnail =
-                        Image::fromUrl({root.value("thumbnail").toString()});
                     if (getSettings()->unshortLinks)
                     {
                         linkString = root.value("link").toString();