+++ src/providers/twitch/IrcMessageHandler.cpp
@@ -1,4 +1,4 @@
-﻿#include "IrcMessageHandler.hpp"
+#include "IrcMessageHandler.hpp"
 
 #include "Application.hpp"
 #include "controllers/accounts/AccountController.hpp"
@@ -377,7 +377,8 @@ void IrcMessageHandler::handleClearChatMessage(Communi::IrcMessage *message)
     {
         chan->disableAllMessages();
         chan->addMessage(
-            makeSystemMessage("Chat has been cleared by a moderator."));
+            makeSystemMessage("Chat has been cleared by a moderator.",
+                              calculateMessageTimestamp(message)));
 
         return;
     }
@@ -397,9 +398,10 @@ void IrcMessageHandler::handleClearChatMessage(Communi::IrcMessage *message)
         reason = v.toString();
     }
 
-    auto timeoutMsg = MessageBuilder(timeoutMessage, username,
-                                     durationInSeconds, reason, false)
-                          .release();
+    auto timeoutMsg =
+        MessageBuilder(timeoutMessage, username, durationInSeconds, reason,
+                       false, calculateMessageTimestamp(message))
+            .release();
     chan->addOrReplaceTimeout(timeoutMsg);
 
     // refresh all
@@ -563,8 +565,9 @@ std::vector<MessagePtr> IrcMessageHandler::parseUserNoticeMessage(
 
     if (it != tags.end())
     {
-        auto b = MessageBuilder(systemMessage,
-                                parseTagString(it.value().toString()));
+        auto b =
+            MessageBuilder(systemMessage, parseTagString(it.value().toString()),
+                           calculateMessageTimestamp(message));
 
         b->flags.set(MessageFlag::Subscription);
         auto newMessage = b.release();
@@ -604,8 +607,9 @@ void IrcMessageHandler::handleUserNoticeMessage(Communi::IrcMessage *message,
 
     if (it != tags.end())
     {
-        auto b = MessageBuilder(systemMessage,
-                                parseTagString(it.value().toString()));
+        auto b =
+            MessageBuilder(systemMessage, parseTagString(it.value().toString()),
+                           calculateMessageTimestamp(message));
 
         b->flags.set(MessageFlag::Subscription);
         auto newMessage = b.release();
@@ -658,33 +662,32 @@ std::vector<MessagePtr> IrcMessageHandler::parseNoticeMessage(
 {
     if (message->content().startsWith("Login auth", Qt::CaseInsensitive))
     {
-        return {MessageBuilder(systemMessage,
-                               "Login expired! Try logging in again.")
-                    .release()};
+        const auto linkColor = MessageColor(MessageColor::Link);
+        const auto accountsLink = Link(Link::OpenAccountsPage, QString());
+        const auto curUser = getApp()->accounts->twitch.getCurrent();
+        const auto expirationText = QString("Login expired for user \"%1\"!")
+                                        .arg(curUser->getUserName());
+        const auto loginPromptText = QString(" Try adding your account again.");
+
+        auto builder = MessageBuilder();
+        builder.message().flags.set(MessageFlag::System);
+
+        builder.emplace<TimestampElement>();
+        builder.emplace<TextElement>(expirationText, MessageElementFlag::Text,
+                                     MessageColor::System);
+        builder
+            .emplace<TextElement>(loginPromptText, MessageElementFlag::Text,
+                                  linkColor)
+            ->setLink(accountsLink);
+
+        return {builder.release()};
     }
     else
     {
         std::vector<MessagePtr> builtMessages;
 
-        if (message->tags().contains("historical"))
-        {
-            bool customReceived = false;
-            qint64 ts = message->tags()
-                            .value("rm-received-ts")
-                            .toLongLong(&customReceived);
-            if (!customReceived)
-            {
-                ts = message->tags().value("tmi-sent-ts").toLongLong();
-            }
-
-            QDateTime dateTime = QDateTime::fromMSecsSinceEpoch(ts);
-            builtMessages.emplace_back(
-                makeSystemMessage(message->content(), dateTime.time()));
-        }
-        else
-        {
-            builtMessages.emplace_back(makeSystemMessage(message->content()));
-        }
+        builtMessages.emplace_back(makeSystemMessage(
+            message->content(), calculateMessageTimestamp(message)));
 
         return builtMessages;
     }