+++ src/widgets/settingspages/KeyboardSettingsPage.cpp
@@ -66,6 +66,10 @@ KeyboardSettingsPage::KeyboardSettingsPage()
     form->addRow(new QLabel("Ctrl + P"), new QLabel("Open Settings menu"));
     form->addRow(new QLabel("F5"),
                  new QLabel("Reload subscriber and channel emotes"));
+
+    form->addItem(new QSpacerItem(16, 16));
+    form->addRow(new QLabel("PageUp"), new QLabel("Scroll up"));
+    form->addRow(new QLabel("PageDown"), new QLabel("Scroll down"));
 }
 
 }  // namespace chatterino